# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loom API",
    "description": "Subtitle generation engine over loom_core.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/romanize": {
      "post": {
        "tags": [
          "text"
        ],
        "summary": "Romanize",
        "operationId": "romanize_romanize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RomanizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RomanizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/romanize/batch": {
      "post": {
        "tags": [
          "text"
        ],
        "summary": "Romanize Batch",
        "operationId": "romanize_batch_romanize_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RomanizeBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RomanizeBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RomanizeBatchItem": {
        "properties": {
          "romanized": {
            "type": "string",
            "title": "Romanized"
          }
        },
        "type": "object",
        "required": [
          "romanized"
        ],
        "title": "RomanizeBatchItem",
        "description": "One result entry \u2014 just the romanized string.  Per-call\nmetadata (lang_code, romanization_name, has_phonetic_layer) is\nconstant across the batch and lives at the response root."
      },
      "RomanizeBatchRequest": {
        "properties": {
          "texts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 2000,
            "title": "Texts",
            "description": "UTF-8 source texts to romanize.  All texts share the lang/phonetic_system/long_vowel_mode specified at the request level.  Hard caps: \u22642000 entries, each \u22645000 chars."
          },
          "lang_code": {
            "type": "string",
            "title": "Lang Code",
            "description": "See POST /romanize."
          },
          "phonetic_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phonetic System",
            "description": "See POST /romanize."
          },
          "long_vowel_mode": {
            "type": "string",
            "title": "Long Vowel Mode",
            "description": "See POST /romanize.",
            "default": "macrons"
          },
          "opt_in_training": {
            "type": "boolean",
            "title": "Opt In Training",
            "description": "See POST /romanize.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "texts",
          "lang_code"
        ],
        "title": "RomanizeBatchRequest"
      },
      "RomanizeBatchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/RomanizeBatchItem"
            },
            "type": "array",
            "title": "Results",
            "description": "One entry per input text, same order as the request.  Empty/oversized texts and unsupported languages produce {romanized: ''} instead of being dropped, so positional alignment with the request is preserved."
          },
          "lang_code": {
            "type": "string",
            "title": "Lang Code"
          },
          "romanization_name": {
            "type": "string",
            "title": "Romanization Name"
          },
          "has_phonetic_layer": {
            "type": "boolean",
            "title": "Has Phonetic Layer"
          }
        },
        "type": "object",
        "required": [
          "results",
          "lang_code",
          "romanization_name",
          "has_phonetic_layer"
        ],
        "title": "RomanizeBatchResponse"
      },
      "RomanizeRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 5000,
            "title": "Text",
            "description": "UTF-8 source text to romanize (\u22645000 chars)."
          },
          "lang_code": {
            "type": "string",
            "title": "Lang Code",
            "description": "BCP-47 language tag (ja, zh-Hans, zh-Hant, yue, ko, th, ru, hi, he, ar, fa, ur, ...)."
          },
          "phonetic_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phonetic System",
            "description": "Per-language phonetic-system override.  Thai: paiboon|rtgs|ipa.  Arabic: learner|din|loose.  Persian: learner|dmg.  Urdu: learner|ala-lc.  Chinese: pinyin|zhuyin|jyutping (typically auto-derived from lang_code).  When omitted, falls back to the language's default."
          },
          "long_vowel_mode": {
            "type": "string",
            "title": "Long Vowel Mode",
            "description": "Japanese-only.  One of macrons|doubled|unmarked.  Ignored for other languages.",
            "default": "macrons"
          },
          "opt_in_training": {
            "type": "boolean",
            "title": "Opt In Training",
            "description": "When true, the (text, lang, romanized) tuple may be archived for OCR training data (Step 5+).  No-op until the pipeline lands.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "text",
          "lang_code"
        ],
        "title": "RomanizeRequest"
      },
      "RomanizeResponse": {
        "properties": {
          "romanized": {
            "type": "string",
            "title": "Romanized"
          },
          "lang_code": {
            "type": "string",
            "title": "Lang Code"
          },
          "romanization_name": {
            "type": "string",
            "title": "Romanization Name"
          },
          "has_phonetic_layer": {
            "type": "boolean",
            "title": "Has Phonetic Layer"
          }
        },
        "type": "object",
        "required": [
          "romanized",
          "lang_code",
          "romanization_name",
          "has_phonetic_layer"
        ],
        "title": "RomanizeResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}